LayoutTests/platform/ios-sim/fast/dom/HTMLMeterElement/meter-clone-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'null' is not an object (evaluating 'targetShadowRoot.firstChild')
PASS cloned.value is target.value
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
